Rename step1 package
