+++ src/test/kotlin/person/PersonTest.kt
@@ -1,4 +1,4 @@
-package step1
+package person
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
